+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.2.4 (2016-10-09)
+
+### Testing
+
+* Add Karma config
+* Add Travis-CI config
+
 # 0.2.3 (2016-10-04)
 
 ### Features